git-svn-id: svn+ssh://ginandtonic.no/home/kms/.svnrepos/scout-kenwood@92 bec0a6a6-59ba-0310-824c-86e0bd1f4ed1

+++ fRound.h
@@ -11,7 +11,7 @@
 
 typedef struct {
     uint32_t lower;
-    uint32_t step;
+    uint8_t step;
 } freq_bands;
 
 uint32_t roundFreq(uint32_t freq);
